Changes to be committed: 	modified:   port/builtins/modaudio.c 	modified:   port/drivers/audio/include/recorder.h 	modified:   port/drivers/audio/recorder.c 	modified:   port/drivers/audio/wav_codec.c 	new file:   port/test/2.wav 	new file:   port/test/3.wav 	new file:   port/test/4.wav 	new file:   port/test/boot.py 	new file:   port/test/pluck-pcm16.wav 	new file:   port/test/test.wav

+++ port/drivers/audio/include/recorder.h
@@ -12,7 +12,7 @@ typedef struct
     int frame_size;
     const char *file_uri;
     uint8_t time;
-    uint8_t total_frames;
+    uint16_t total_frames;
     EventGroupHandle_t recorder_event;
     RingbufHandle_t stream_in_ringbuff;
 } recorder_handle_t;
